Roulette Logic
I've implemented the roulette logic into the game and rand some tests, took a while to get it working but did it in the end. now the roulette button unlocks a sprite button for the player. Next is to get the buttons themselves to change the background and sprite colours and once thats done its all about making it look showy. i.e. design make the roulette more exciting, so players truly believe the roulette is what determines their outcome not the button that takes them to the roulette screen.