colour update 2
ok so tried making the particles update got it half right. x is not updating properly but logic is working so need to go back and tweak that. once complete i just need to set up all the colours with the apropriate shades. and then make it so player prefs keeps the unlocked skins unlocked. concurrently the token counter appears to reset when going from one scene to another. if i play the program stop it and rejoin it works fine but for some reason going between scenes isnt keeping the token counter and tokens up to date with the player prefs...