Colour Fix
had to make it so the script loaded the new colours on start in both scenes in order to have the colour update stay across multiple scenes. also fixed the particle colours by also adding that into the update at the start. I think its a similar issue regarding the tokens so ill need to look into that, so that the timer does not freeze or reset when swapping scenes in the game...